refactor(daemon): chanage package path
